Reduce duplication via async-only pipelines
